virtual populate and api improvements

+++ controllers/diagnosisController.js
@@ -1,10 +1,20 @@
 const catchAsync = require('../utils/catchAsync')
 const AppError = require('../utils/appError')
 const Diagnosis = require('../models/DiagnosisModel')
+const Pacient = require('../models/PacientModel')
+const Employee = require('../models/EmployeeModel')
 const cloudinary = require('cloudinary').v2
 
 exports.createDiagnosis = catchAsync(async (req, res, next) => {
     const { pacientId } = req.params
+    // WORKAROUND
+    const pacient = await Pacient.findOne({ _id: pacientId })
+    const employee = await Employee.findOne({ _id: req.user._id })
+
+    if (!pacient) {
+        return next(new AppError('Dogodila se greška pri dodavanju dijagnoze za ovog pacijenta.', 404))
+    }
+    //////////
 
     const newDiagnosis = await Diagnosis.create({
         pacientId,
@@ -19,7 +29,9 @@ exports.createDiagnosis = catchAsync(async (req, res, next) => {
 
     res.status(201).json({
         message: 'success',
-        newDiagnosis
+        newDiagnosis,
+        pacientName: `${pacient.firstName} ${pacient.lastName}`,
+        employeeName: `${employee.firstName} ${employee.lastName}`,
     })
 })
 
@@ -39,24 +51,10 @@ exports.getAllDiagnosis = catchAsync(async (req, res, next) => {
 
 exports.getOneDiagnosis = catchAsync(async (req, res, next) => {
     const { diagnosisId } = req.params
-    const diagnosis = await Diagnosis.find({ _id: diagnosisId })
+    const diagnosis = await Diagnosis.findOne({ _id: diagnosisId })
 
     if (!diagnosis) {
         return next(new AppError('Dijagnoza nije pronađena', 404))
-    }
-
-    res.status(200).json({
-        message: 'success',
-        diagnosis
-    })
-})
-
-exports.getPacientDiagnosis = catchAsync(async (req, res, next) => {
-    const { pacientId } = req.params
-    const diagnosis = await Diagnosis.find({ pacientId })
-
-    if (!diagnosis) {
-        return next(new AppError('Nijedna dijagnoza nije pronađena za ovog pacijenta.', 404))
     }
 
     res.status(200).json({
@@ -74,7 +72,7 @@ exports.updateDiagnosis = catchAsync(async (req, res, next) => {
     }
 
     diagnosis.pacientId = req.body.pacientId || diagnosis.pacientId
-    diagnosis.employeeId = req.body.employeeId || diagnosis.employeeId
+    diagnosis.employeeId = req.body.employeeId || diagnosis.employeeId._id
     diagnosis.date = req.body.date || diagnosis.date
     diagnosis.summary = req.body.summary || diagnosis.summary
     diagnosis.image = req.files ? req.files.image : diagnosis.image
@@ -93,7 +91,7 @@ exports.updateDiagnosis = catchAsync(async (req, res, next) => {
 exports.deleteDiagnosis = catchAsync(async (req, res, next) => {
     const { diagnosisId } = req.params
     const diagnosis = await Diagnosis.findOne({ _id: diagnosisId })
-    const publicId = diagnosis.image.split('/')[7].split('.')[0]
+    let publicId
 
     if (!diagnosis) {
         return next(new AppError('Dijagnozu nije moguće obrisati. Za pomoć kontaktirajte korisničku podršku.', 404))
@@ -102,12 +100,16 @@ exports.deleteDiagnosis = catchAsync(async (req, res, next) => {
     try {
         await Diagnosis.findByIdAndDelete(diagnosisId)
 
-        await cloudinary.api.delete_resources(publicId, { invalidate: true },
-            function (error, result) {
-                if (error) {
-                    console.log(error)
-                }
-            });
+        if (diagnosis.image) {
+            publicId = diagnosis.image.split('/')[7].split('.')[0]
+
+            await cloudinary.api.delete_resources(publicId, { invalidate: true },
+                function (error, result) {
+                    if (error) {
+                        console.log(error)
+                    }
+                });
+        }
     } catch (err) {
         console.log(err)
     }
